Add potions into the game. Refactor Attributes.
Potions are not generated yet, but six types of potions are supported.
Health potions apply their effect directly on the player, and stat
potions apply a decorator on the player's attributes.

Attributes is changed to implement the AttributeProvider interface. The
AttributeDecorator class is added to represent a decorator on an
AttributeProvider. The PotionBuff class is added as a subclass of
AttributeDecorator which provides the ATK/DEF buff/debuff applied by
potions.

Headers that previously included levelobjectvisitor now do not. Instead,
the implementation of the `accept` methods is moved into the
implementation files.