Add field of view.
The Memory class stores a grid of the most recently seen version of the
actual level. It is updated by Level::drawPOV, which, given the player's
position, computes the field of view and also draws everything visible
in that field onto the screen.

To make this happen, Level no longer adds things onto the Display.
I don't think it makes use of the Display object at all, actually.
Instead it draws items individually onto UI.

I needed to change the color_pair calculation to add 1, otherwise
black/black became 0, which is not allowed.

The code isn't very pretty right now. It's also not switchable on/off,
as it currently stands. Probably Game should know whether or not FOV is
on, and if it isn't, it shouldn't create a Memory object at all, but
instead call a different method on Level that draws the whole thing.

Lighting will come later. This simply involves an additional check in
Level::drawPOV to see if the tiles in the field of view are sufficiently
lit.