Fix FOV to have more natural shadows.
This makes the algorithm simultaneously more restrictive and more
permissive, by allowing more cells to be lit but disallowing
disconnected cells. This ensures that there are no strange lit regions
far away from any other lit cell.

Note that the reference implementation is completely insane, so I don't
do the same thing. It is slightly more permissive, but I believe that's
because of a bug.